requirements.txt
click==8.1.3
PyQt5==5.15.6
PyYAML==6.0
SQLAlchemy==1.4.36
